// src/app/shared/models/travelling.model.ts
/*Authored by Areeba
HR Travelling SCR
Dated - 07-06-2022
*/

export class Travelling {
    hrTrvlngSeq: number;
    refCdTrvlngRol: number;
    refCdCalcTyp: number;
    refCdPortTypSeq: number;
    fromDt: Date;
    toDt: Date;
    amt: number;

    rol: String;
    portTyp: string;
    calcTyp: string;
}